remove non-root user
Using a non-root user left me unable (sometimes, don't know why) to
rename the utility bill files in the location they were mounted. This
was only a problem on Linux, but worth it to solve in order to be
cross-platform compatible.

Now the danger is simply that whatever directory is mounted at /data is
subject to the docker container user.