make move command verbose in place of echo statement
Originally, I ran an echo command to show what was about to be done.
However, the next line checks if the previous command executed (which it
always would, because it's an echo) and that was not my intention.

Now it checks if the previous *python* command executed successfully,
and moves the file verbosely if it did.