Feat: Add config to address PDF versions.
add actual element to logs

add elements to log messages

add config file to outer docker layer

Introduce PDF revision config as yaml.

Create a local dotfile config for specifying field numbers (where to
find the date and amount fields) for each revision of the utility bill.
The script gets a rough date from the PDF itself, then finds the latest
version config for which the date of the revision is less than the date
of the bill it is processing.

loop date parsing so failures can be omitted

fix: correct config file extension